template code divide
