+++ lib/features/templates/widgets/tag_chip.dart
@@ -1,0 +1,29 @@
+// Flutter imports:
+import 'package:flutter/material.dart';
+
+class TagChip extends StatelessWidget {
+  const TagChip({super.key, required this.label});
+  final String label;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
+      decoration: BoxDecoration(
+        color: theme.colorScheme.surfaceContainerHighest,
+        borderRadius: BorderRadius.circular(999),
+        border: Border.all(
+          color: theme.colorScheme.outlineVariant.withValues(alpha: 0.15),
+        ),
+      ),
+      child: Text(
+        label,
+        style: theme.textTheme.labelSmall?.copyWith(
+          color: theme.colorScheme.primary.withValues(alpha: 0.8),
+          fontWeight: FontWeight.w500,
+        ),
+      ),
+    );
+  }
+}
