+++ lib/features/templates/widgets/templates_widgets.dart
@@ -1,370 +1,6 @@
-// Flutter imports:
-import 'package:flutter/material.dart';
-
-// Package imports:
-import 'package:line_icons/line_icons.dart';
-
-class TemplatesSearchField extends StatelessWidget {
-  const TemplatesSearchField({super.key, required this.onChanged});
-  final ValueChanged<String> onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16, 12, 16, 8),
-      child: TextField(
-        style: TextStyle(color: theme.colorScheme.primary),
-        decoration: InputDecoration(
-          hintText: 'Search templates...',
-          hintStyle: TextStyle(
-              color: theme.colorScheme.primary.withValues(alpha: 0.6)),
-          prefixIcon: Icon(Icons.search,
-              color: theme.colorScheme.primary.withValues(alpha: 0.8)),
-          filled: true,
-          fillColor: theme.colorScheme.surface.withValues(alpha: 0.9),
-          contentPadding:
-              const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-          isDense: true,
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(12),
-            borderSide: BorderSide(
-              color: theme.colorScheme.outlineVariant.withValues(alpha: 0.15),
-            ),
-          ),
-          enabledBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(12),
-            borderSide: BorderSide(
-              color: theme.colorScheme.outlineVariant.withValues(alpha: 0.15),
-            ),
-          ),
-          focusedBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(12),
-            borderSide:
-                BorderSide(color: theme.colorScheme.primary, width: 1.0),
-          ),
-        ),
-        onChanged: (v) => onChanged(v.trim()),
-      ),
-    );
-  }
-}
-
-class TemplateListItem extends StatelessWidget {
-  const TemplateListItem({
-    super.key,
-    required this.title,
-    required this.tags,
-    required this.onTap,
-    this.onLongPress,
-    required this.onEdit,
-    required this.onDelete,
-    this.isSelected = false,
-  });
-  final String title;
-  final List<String> tags;
-  final VoidCallback onTap;
-  final VoidCallback? onLongPress;
-  final VoidCallback onEdit;
-  final VoidCallback onDelete;
-  final bool isSelected;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 6.0),
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          borderRadius: BorderRadius.circular(12),
-          onTap: onTap,
-          onLongPress: onLongPress,
-          child: Container(
-            padding: const EdgeInsets.all(16),
-            decoration: BoxDecoration(
-              color: isSelected
-                  ? theme.colorScheme.primary.withValues(alpha: 0.1)
-                  : theme.colorScheme.surface.withValues(alpha: 0.3),
-              borderRadius: BorderRadius.circular(12),
-              border: Border.all(
-                color: isSelected
-                    ? theme.colorScheme.primary.withValues(alpha: 0.3)
-                    : theme.colorScheme.outline.withValues(alpha: 0.1),
-              ),
-              boxShadow: [
-                BoxShadow(
-                  color: theme.colorScheme.shadow.withValues(alpha: 0.05),
-                  blurRadius: 8,
-                  offset: const Offset(0, 2),
-                )
-              ],
-            ),
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                Container(
-                  width: 48,
-                  height: 48,
-                  decoration: BoxDecoration(
-                    color: theme.colorScheme.primary.withValues(alpha: 0.1),
-                    borderRadius: BorderRadius.circular(12),
-                  ),
-                  child: Icon(
-                    LineIcons.fileAlt,
-                    color: theme.colorScheme.primary,
-                    size: 24,
-                  ),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        title,
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                        style: theme.textTheme.titleMedium?.copyWith(
-                          color: theme.colorScheme.onSurface,
-                          fontWeight: FontWeight.w600,
-                        ),
-                      ),
-                      if (tags.isNotEmpty) ...[
-                        const SizedBox(height: 4),
-                        Wrap(
-                          spacing: 6,
-                          runSpacing: 4,
-                          children:
-                              tags.map((t) => _TagChip(label: t)).toList(),
-                        ),
-                      ],
-                    ],
-                  ),
-                ),
-                if (isSelected)
-                  Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      OutlinedButton.icon(
-                        onPressed: onEdit,
-                        icon: const Icon(Icons.edit, size: 18),
-                        label: const Text('Edit'),
-                        style: OutlinedButton.styleFrom(
-                          foregroundColor: theme.colorScheme.primary,
-                          side: BorderSide(
-                            color: theme.colorScheme.outline
-                                .withValues(alpha: 0.2),
-                          ),
-                          padding: const EdgeInsets.symmetric(
-                              horizontal: 12, vertical: 8),
-                          textStyle: theme.textTheme.labelMedium,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(8),
-                          ),
-                        ),
-                      ),
-                      const SizedBox(width: 8),
-                      OutlinedButton.icon(
-                        onPressed: onDelete,
-                        icon: const Icon(Icons.delete_outline, size: 18),
-                        label: const Text('Delete'),
-                        style: OutlinedButton.styleFrom(
-                          foregroundColor: theme.colorScheme.error,
-                          side: BorderSide(
-                            color:
-                                theme.colorScheme.error.withValues(alpha: 0.3),
-                          ),
-                          padding: const EdgeInsets.symmetric(
-                              horizontal: 12, vertical: 8),
-                          textStyle: theme.textTheme.labelMedium,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(8),
-                          ),
-                        ),
-                      ),
-                    ],
-                  )
-                else
-                  Icon(
-                    Icons.arrow_forward_ios,
-                    size: 16,
-                    color: theme.colorScheme.onSurface.withValues(alpha: 0.5),
-                  ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _TagChip extends StatelessWidget {
-  const _TagChip({required this.label});
-  final String label;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
-      decoration: BoxDecoration(
-        color: theme.colorScheme.surfaceContainerHighest,
-        borderRadius: BorderRadius.circular(999),
-        border: Border.all(
-          color: theme.colorScheme.outlineVariant.withValues(alpha: 0.15),
-        ),
-      ),
-      child: Text(
-        label,
-        style: theme.textTheme.labelSmall?.copyWith(
-          color: theme.colorScheme.primary.withValues(alpha: 0.8),
-          fontWeight: FontWeight.w500,
-        ),
-      ),
-    );
-  }
-}
-
-class TagChipsRow extends StatelessWidget {
-  const TagChipsRow({super.key, required this.tags, required this.onRemove});
-  final List<String> tags;
-  final ValueChanged<String> onRemove;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    if (tags.isEmpty) return const SizedBox.shrink();
-    return SizedBox(
-      height: 30,
-      child: ListView.builder(
-        scrollDirection: Axis.horizontal,
-        itemCount: tags.length,
-        itemBuilder: (context, index) {
-          final tag = tags[index];
-          return Container(
-            margin: const EdgeInsets.only(right: 6),
-            child: Chip(
-              label: Text(
-                tag,
-                style: TextStyle(
-                  fontSize: 13,
-                  color: theme.colorScheme.primary.withValues(alpha: 0.85),
-                  fontWeight: FontWeight.w500,
-                ),
-              ),
-              backgroundColor: theme.colorScheme.surfaceContainerHighest,
-              deleteIcon: Icon(
-                Icons.close,
-                size: 18,
-                color: theme.colorScheme.primary.withValues(alpha: 0.75),
-              ),
-              onDeleted: () => onRemove(tag),
-              materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-              visualDensity: VisualDensity.compact,
-              padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-              shape: StadiumBorder(
-                side: BorderSide(
-                  color:
-                      theme.colorScheme.outlineVariant.withValues(alpha: 0.15),
-                ),
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
-
-class TagInputField extends StatelessWidget {
-  const TagInputField({
-    super.key,
-    required this.controller,
-    required this.focusNode,
-    required this.onSubmit,
-  });
-  final TextEditingController controller;
-  final FocusNode focusNode;
-  final ValueChanged<String> onSubmit;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return SizedBox(
-      height: 40,
-      child: Row(
-        children: [
-          Expanded(
-            child: TextField(
-              controller: controller,
-              focusNode: focusNode,
-              textInputAction: TextInputAction.done,
-              onSubmitted: (raw) {
-                final value = raw.trim();
-                if (value.isEmpty) return;
-                onSubmit(value);
-                controller.clear();
-                FocusScope.of(context).unfocus();
-              },
-              style: const TextStyle(fontSize: 14),
-              decoration: InputDecoration(
-                hintText: 'Add tag and press +',
-                hintStyle: TextStyle(
-                  fontSize: 12,
-                  color: theme.colorScheme.primary.withValues(alpha: 0.5),
-                ),
-                prefixIcon: Icon(
-                  Icons.tag,
-                  size: 16,
-                  color: theme.colorScheme.primary.withValues(alpha: 0.7),
-                ),
-                suffixIcon: IconButton(
-                  icon: Icon(Icons.add,
-                      size: 18,
-                      color: theme.colorScheme.primary.withValues(alpha: 0.8)),
-                  tooltip: 'Add tag',
-                  onPressed: () {
-                    final value = controller.text.trim();
-                    if (value.isEmpty) return;
-                    onSubmit(value);
-                    controller.clear();
-                    FocusScope.of(context).unfocus();
-                  },
-                  padding: EdgeInsets.zero,
-                  constraints:
-                      const BoxConstraints(minWidth: 32, minHeight: 32),
-                ),
-                filled: true,
-                fillColor: theme.colorScheme.surface,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(18),
-                  borderSide: BorderSide(
-                    color: theme.colorScheme.outlineVariant
-                        .withValues(alpha: 0.15),
-                  ),
-                ),
-                enabledBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(18),
-                  borderSide: BorderSide(
-                    color: theme.colorScheme.outlineVariant
-                        .withValues(alpha: 0.15),
-                  ),
-                ),
-                focusedBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(18),
-                  borderSide:
-                      BorderSide(color: theme.colorScheme.primary, width: 1.0),
-                ),
-                contentPadding:
-                    const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-                isDense: true,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
+// Export all template widgets for easy importing
+export 'templates_search_field.dart';
+export 'template_list_item.dart';
+export 'tag_chip.dart';
+export 'tag_chips_row.dart';
+export 'tag_input_field.dart';
